Remove checkpoints in ASA: Getting to the device
This means the player now must win this mission in one go.

Justification: There were many ways for the mission to be saved in an unwinnable state,
there are many ways to win this mission and the checkpoints try to "force" one
particular way. Also, this mission isn't too long anyway.