Remove 2 unused number tags in Construction Mode GUI
These numbers are shown aside the power tag, but the numbers never change.
They don't serve any purpose and are just visual clutter and annoying, since
they partially overlap.
They are probably a leftover from copying code over from other scripts.

With this changeset, only the power and turn time are left visible, as it is
supposed to.