Racer: Fix weird water splashes after waypoint placement
Does not affect official racer, as only waypoint placement is touched.

The reason was that the air attack gear sometimes was not deleted fast enough so it might occassionally drop some air bombs (these are deleted now). Also, the airplane position was set to water level, which caused another water splash.