Allow to see rooms of incompatible versions in the lobby
For the new clients the room version is shown in a separate column.

There is also a hack for previous versions clients: the room vesion
specifier is prepended to the room names for rooms of incompatible versions,
and the server shows 'incompatible version' error if the client tries to join them.